bet_price: keep info_for_display and applicable_ticks in sync
It seems like we are getting extra ticks (which are renumbered to the
"start" of the contract because the signal used to know when to stop
displaying (applicable_ticks) is not kept in sync with the actual source
of the data.

This doesn't _seem_ quite right, but it may be a problem in a small,
racy corner case.  It's essentially impossible to reproduce on demand
and testing is even more difficult, so we'll assume this is good enough
until it is not.